+++ routes/users_routes.js
@@ -4,7 +4,9 @@ const router = express.Router()
 //const checkAuth = auth_middleware.checkAuth
 const axios = require('axios')
 
-const { validateApiKey } = require("../middleware/api_key_validation.js");
+const { validateApiKey, serviceIsActive } = require("../middleware/api_key_validation.js");
+
+const serviceName = 'Users';
 
 let url_users = process.env.URL_USERS;
 
@@ -20,38 +22,45 @@ router.all('*', async function(req, res) {
         res.json({message: "Invalid API Key"});
         return;
     } else {
-        let url = url_users + req.originalUrl
-    let method = req.method
-    console.log("[PROXY " + method + "]:", url)
-    
-    let axios_promise
-
-    switch(method) {
-        case "GET":
-            axios_promise = axios.get(url, req.body)
-            break
-        case "POST":
-            axios_promise = axios.post(url, req.body)
-            break
-        case "PUT":
-            axios_promise = axios.put(url, req.body)
-            break
-        case "DELETE":
-            axios_promise = axios.delete(url, req.body)
-            break
-    }
-    
-    axios_promise
-    .then(response => {
-        res.statusCode = response.status;
-        res.json({message: response.data})
-    })
-    .catch(error => {
-        res.statusCode = error.response.status;
-        res.json({message: error.response.data})
-    });
-    }
-    
+        const serviceActive = await serviceIsActive(serviceName);
+        if (!serviceActive){
+            res.statusCode = 407;
+            res.json({message: "Users service is currently unavailable."});
+            return;
+        } else {
+            let url = url_users + req.originalUrl
+            let method = req.method
+            console.log("[PROXY " + method + "]:", url)
+            
+            let axios_promise
+        
+            switch(method) {
+                case "GET":
+                    axios_promise = axios.get(url, req.body)
+                    break
+                case "POST":
+                    axios_promise = axios.post(url, req.body)
+                    break
+                case "PUT":
+                    axios_promise = axios.put(url, req.body)
+                    break
+                case "DELETE":
+                    axios_promise = axios.delete(url, req.body)
+                    break
+            }
+            
+            axios_promise
+            .then(response => {
+                res.statusCode = response.status;
+                res.json({message: response.data})
+            })
+            .catch(error => {
+                res.statusCode = error.response.status;
+                res.json({message: error.response.data})
+            });
+            
+         }
+    }         
 })
 
 module.exports = {router}
